Bug fix for multiple results for tasks with similar names (#10)
* Fix non-unique queries returning multiple results.
The answer wasn't to use the task_id (mainly due to the way the RTM results are structured), but to continue using the taskseries.name and using the RtmAPI iterators to only return the 1st result (i.e. only 1 .next() ). This is not 100% unique (e.g. tasks with identical names), but that's ok, since the point of this utility is to return 1 'random' result.

* Update Changelog.